requirements.txt
crewai==0.11.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
google-search-results==2.4.2
langchain==0.1.14
langchain-google-genai==1.0.1
streamlit==1.32.2
unstructured>=0.1.0
